Moving enums to dto package
