+++ data/src/main/java/io/gnosis/data/adapters/OperationEnumAdapter.kt
@@ -3,7 +3,7 @@ package io.gnosis.data.adapters
 import com.squareup.moshi.FromJson
 import com.squareup.moshi.JsonDataException
 import com.squareup.moshi.ToJson
-import io.gnosis.data.models.Operation
+import io.gnosis.data.backend.dto.Operation
 
 class OperationEnumAdapter {
     @ToJson
