+++ data/src/main/java/io/gnosis/data/models/Transaction.kt
@@ -29,33 +29,3 @@ sealed class Transaction {
     ) : Transaction()
 }
 
-enum class ContractInfoType {
-    ERC20, // default
-    ERC721
-}
-
-enum class Operation(val id: Int) {
-    CALL(0),
-    DELEGATE(1)
-}
-
-enum class TransferType {
-    ETHER_TRANSFER,
-    ERC20_TRANSFER,
-    ERC721_TRANSFER,
-    UNKNOWN
-}
-
-enum class TransactionType {
-    ETHEREUM_TRANSACTION,
-    MULTISIG_TRANSACTION,
-    MODULE_TRANSACTION,
-    UNKNOWN
-}
-
-enum class SignatureType {
-    CONTRACT_SIGNATURE,
-    APPROVED_HASH,
-    EOA,
-    ETH_SIGN
-}
